Create a more pay-phone-esque 'parental control'
The service is supposed to monitor
1. The amount of time a user is logged in
2. The amount of time the user spends on a particular program.
3. The amount of time the user has a 'drive' mounted to the system.

The program must then be controllable to the point of 1. logging out of
the user profile 2. shut down the computer 3. eject the medium.

I have no idea how to do any of these thing, nor do I know the first
thing about visual C#. So this ought to be a treat :+1:.